+++ src/Components/Dash/Contracts.js
@@ -2,7 +2,7 @@ export const tokenContracts = [
     { address: '0x1628160C66e0330090248a163A34Ba5B5A82D4f7', symbol: 'DogSwap', decimals: 12, image: '/images/coins/dogswap.png', homepage: 'https://dogswap.xyz', info: 'DogSwap is a decentralized exchange for tokens created on the MintMe smartchain.', url: 'https://mintme.com/token/DogSwap' },
     { address: '0x9D8dd79F2d4ba9E1C3820d7659A5F5D2FA1C22eF', symbol: 'BONE', decimals: 18, image: '/images/coins/bone.png', homepage: 'https://dogswap.xyz', info: 'BONE is the governance token for the DogSwap ecosystem. Stake it, Earn it', url: 'https://mintme.com/token/BONE' },
     { address: '0x7b535379bBAfD9cD12b35D91aDdAbF617Df902B2', symbol: '1000x', decimals: 12, image: '/images/coins/1000x.png', homepage: 'https://1000x.ch', info: '1000x is one of the first tokens on the MintMe smartchain and one of the leading projects backing DogSwap.', url: 'https://mintme.com/token/1000x' },
-    { address: '0xB580f1dbA1c17882Fca8f6DDadA8428c9cB177fC', symbol: 'XatteR', decimals: 12, image: '/images/coins/xatter.jpeg', homepage: 'https://AncientBeast.com', info: 'Ancient Beast is a turn based strategy game, eSport played in 1vs1 or 2vs2 matches', url: 'https://mintme.com/token/XatteR' }
+    { address: '0xB580f1dbA1c17882Fca8f6DDadA8428c9cB177fC', symbol: 'XatteR', decimals: 12, image: '/images/coins/xatter.png', homepage: 'https://AncientBeast.com', info: 'Ancient Beast is a turn based strategy game, eSport played in 1vs1 or 2vs2 matches', url: 'https://mintme.com/token/XatteR' }
   ];
   
   export const routerAddress = '0xa6c72da53025c5291d326fadc5a1277aa21fef2c';