+++ src/constants/featuredPool.js
@@ -17,7 +17,7 @@ export const poolData = [
         poolType: "Community",
         lpTokenAddress: "0x1d2a64b660E572ce653f35Ce5A9A655D05ae9fd0",
         imageTokenA: "https://dogswap.xyz/images/coins/mintme.png",
-        imageTokenB: "https://dogswap.xyz/images/coins/xatter.jpeg",
+        imageTokenB: "https://dogswap.xyz/images/coins/xatter.png",
         projectLink: "https://ancientbeast.com"
     },
 ];