+++ CHANGELOG.md
@@ -5,6 +5,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.2.7] - 2024-10-17
+
+### Changes
+
+- Split Staking page in components for optimal coding (view is still same for now as i'm bugfixing)
+- XatteR logo rebranding
+
+### Bugfixes
+
+- Collection button has been optimised
+
+### Additions
+
+- You can now add tokens to your wallet by pressing the + sign next to the tokens on the swap popup modal
+(Tokens that are over 12 characters will give you an error message)
+
 ## [2.2.6] - 2024-10-17
 
 ### Changes